chore: move update logic to manager
